requirements.txt
Django==1.10.1
django-background-tasks==1.0.14
django-bootstrap3==7.1.0
django-compat==1.0.13
django-mongoengine==0.2.1
django-pagination==1.0.7
mongoengine==0.10.6
Pillow==3.3.1
pkg-resources==0.0.0
pymongo==3.3.0
six==1.10.0
Unipath==1.1
